requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langserve
python-dotenv
